Split off ExpressionBuilder into separate file for reusability. Testing better ways to allocate space to collapsible panels
